requirements.txt
python-pptx==1.0.2
Pillow==11.1.0
PyMuPDF==1.25.4
pyinstaller==6.12.0
